Move generic Avro code into baseavro.pyx
This means that other modules, such as the upcoming rsdos reader,
don't need to duplicate a ton of code.

Changed flowtuple behaviour slightly to only maintain a single
AvroFlowtuple instance rather than instantiating a fresh instance
everytime we want to read from the file. This matters a lot more
now that the base AvroRecord object does mallocs and frees for
every instance, so it is much more efficient to just have one
instance and reuse it.